feat: 모델 changeDirection 추가 및 파일 변경
